# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: KisanSetu FastAPI contract
  version: 1.0.0
  description: >-
    Browser clients retain same-origin /api paths through the Next.js proxy.
    Numeric decisions, feasibility, and rankings are deterministic Python rules.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/pools/suggestions:
    get:
      summary: Explain compatible and rejected nearby lots for an anchor lot
      parameters:
        - { name: lotId, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Deterministic Geo-Pool suggestions }
...
  /api/pools/{poolId}/join:
    post:
      summary: Revalidate and accept a lot into an existing pool
      parameters:
        - { name: poolId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Persisted pool membership }
...